[sphinx] Integer type should be BigInt
Otherwise filtering would be impossible

+++ app/indices/cms_page_index.rb
@@ -2,7 +2,7 @@
 
 ThinkingSphinx::Index.define 'cms/page', with: :real_time do
   indexes :title
-  has :tenant_id, type: :integer
+  has :tenant_id, type: :bigint
 
   indexes :published
   scope { CMS::Page.where(searchable: true) }
